contact us page, added roles, seo

+++ routes/web.php
@@ -1,6 +1,8 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
+use App\GeneralSetting;
+use App\SocialSetting;
+use App\SeoSetting;
 
 /*
 |--------------------------------------------------------------------------
@@ -38,71 +40,74 @@ Route::get('/offers/thank-you', 'StaticPagesController@thankYou');
 //Admin Dashboard
 Route::get('/admin', 'admin\AdminController@dashboard');
 
+
 //Admin Food Categories
 
-Route::get('/admin/food-categories','admin\FoodCategoriesController@index');
+Route::get('/admin/food-categories','admin\FoodCategoriesController@index')->middleware('role:Admin');
 
-Route::get('/admin/food-categories/create','admin\FoodCategoriesController@create');
+Route::get('/admin/food-categories/create','admin\FoodCategoriesController@create')->middleware('role:Admin');
 
-Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store');
+Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store')->middleware('role:Admin');
 
-Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit');
+Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit')->middleware('role:Admin');
 
-Route::put('/admin/food-categories/{id}', 'admin\FoodCategoriesController@update');
+Route::put('/admin/food-categories/{id}', 'admin\FoodCategoriesController@update')->middleware('role:Admin');
 
-Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete');
+Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete')->middleware('role:Admin');
 
 
 
 //Admin Food Items
 
-Route::get('/admin/food-items','admin\FoodItemsController@index');
+Route::get('/admin/food-items','admin\FoodItemsController@index')->middleware('role:Admin');
 
-Route::get('/admin/food-items/create','admin\FoodItemsController@create');
+Route::get('/admin/food-items/create','admin\FoodItemsController@create')->middleware('role:Admin');
 
-Route::post('/admin/food-items', 'admin\FoodItemsController@store');
+Route::post('/admin/food-items', 'admin\FoodItemsController@store')->middleware('role:Admin');
 
-Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit');
+Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit')->middleware('role:Admin');
 
-Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update');
+Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update')->middleware('role:Admin');
 
-Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete');
+Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete')->middleware('role:Admin');
 
 //Admin Users
 
-Route::get('/admin/users','admin\UsersController@index');
+Route::get('/admin/users','admin\UsersController@index')->middleware('role:Admin');
 
-Route::get('/admin/users/create','admin\UsersController@create');
+Route::get('/admin/users/create','admin\UsersController@create')->middleware('role:Admin');
 
-Route::post('/admin/users','admin\UsersController@store');
+Route::post('/admin/users','admin\UsersController@store')->middleware('role:Admin');
 
-Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit');
+Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit')->middleware('role:Admin');
 
-Route::put('/admin/users/{id}', 'admin\UsersController@update');
+Route::put('/admin/users/{id}', 'admin\UsersController@update')->middleware('role:Admin');
 
-Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete');
+Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete')->middleware('role:Admin');
 
 //Admin Customers
 //Admin Members
 Route::get('/admin/members','admin\MemberController@index');
+
 Route::delete('/admin/members/{id}/delete', 'admin\MemberController@delete');
 
 //Admin Reservations
 Route::get('/admin/reservations','admin\ReservationController@index');
+
 Route::delete('/admin/reservations/{id}/delete', 'admin\ReservationController@delete');
 
 // Admin Settings
-Route::get('/admin/settings/general','admin\SettingController@general');
+Route::get('/admin/settings/general','admin\SettingController@general')->middleware('role:Admin');
 
-Route::put('/admin/settings/general','admin\SettingController@saveGeneral');
+Route::put('/admin/settings/general','admin\SettingController@saveGeneral')->middleware('role:Admin');
 
-Route::get('/admin/settings/seo','admin\SettingController@seo');
+Route::get('/admin/settings/seo','admin\SettingController@seo')->middleware('role:Admin');
 
-Route::put('/admin/settings/seo','admin\SettingController@saveSeo');
+Route::put('/admin/settings/seo','admin\SettingController@saveSeo')->middleware('role:Admin');
 
-Route::get('/admin/settings/social','admin\SettingController@social');
+Route::get('/admin/settings/social','admin\SettingController@social')->middleware('role:Admin');
 
-Route::put('/admin/settings/social','admin\SettingController@saveSocial');
+Route::put('/admin/settings/social','admin\SettingController@saveSocial')->middleware('role:Admin');
 
 
 
@@ -117,6 +122,20 @@ Route::get('/admin/login', function () {
 });
 
 
+
 Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
+
+View::composer(['home', 'pages/about', 'pages/contact', 'pages/offers', 'pages/reservations', 'pages/thank-you', 'menu/all-categories', 'menu/single-menu'], function ($view) {
+    $generalSettings = GeneralSetting::find(1);
+    $socialSettings = SocialSetting::find(1);
+    $seoSettings = SeoSetting::find(1);
+
+    
+    $view->with('settings', [
+        "general" => $generalSettings,
+        "social" => $socialSettings,
+        "seo" => $seoSettings
+    ]);
+});
